feat: implementación inicial de controladores, DTOs, servicios y repositorios
- Se agregaron controladores REST para Accesoria, Socio, Usuario, Producto, Venta, Membresía, Inscripción, Paquete, Entrenador y Categoría
- Se crearon los DTOs asociados organizados por módulo (usuario, membresía, etc.)
- Se implementaron interfaces y clases de servicios con su estructura base
- Se definieron los repositorios JPA para todas las entidades
- Se agregó configuración de ModelMapper (MapperConfig.java)
- Se ajustaron los modelos de entidades para soportar relaciones necesarias y se eliminaron clases obsoletas:
  - Eliminado: ZonaCombate.java, MembresiaServicio.java, TipoServicio.java
  - Modificados: Usuario, Socio, Paquete, Membresía, Entrenador, Categoría
- Se creó la entidad Inscripción y su DTO
- Se actualizó el archivo pom.xml para agregar dependencias necesarias

+++ src/main/java/com/wolfpack/model/Entrenador.java
@@ -17,7 +17,7 @@ public class Entrenador {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @EqualsAndHashCode.Include
-    private Long idEntrenador;
+    private Integer idEntrenador;
 
     @Column(length = 40, nullable = false)
     private String nombre;
